# 10% perf improvement with a little arm64 assembly and laxer limb bounds