--- a/fe_arm64.go
+++ b/fe_arm64.go
@@ -1,8 +1,8 @@
-// Copyright (c) 2019 The Go Authors. All rights reserved.
+// Copyright (c) 2020 The Go Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// +build !amd64,!arm64 !gc purego
+// +build arm64,gc,!purego
 
 package edwards25519
 
@@ -10,6 +10,10 @@ func feMul(v, x, y *fieldElement) { feMulGeneric(v, x, y) }
 
 func feSquare(v, x *fieldElement) { feSquareGeneric(v, x) }
 
+//go:noescape
+func carryPropagate(v *fieldElement)
+
 func (v *fieldElement) carryPropagate() *fieldElement {
-	return v.carryPropagateGeneric()
+	carryPropagate(v)
+	return v
 }