--- a/fe_generic.go
+++ b/fe_generic.go
@@ -101,7 +101,7 @@ func feMulGeneric(v, x, y *fieldElement) {
 	// finally from r_4 to r_0 . Each of these carries is done as one copy, one
 	// right shift by 51, one logical and with 2^51 − 1, and one addition.
 	*v = fieldElement{r00, r10, r20, r30, r40}
-	v.carryPropagate1().carryPropagate2()
+	v.carryPropagate()
 }
 
 func feSquareGeneric(v, x *fieldElement) {
@@ -175,5 +175,5 @@ func feSquareGeneric(v, x *fieldElement) {
 	r00 += r41
 
 	*v = fieldElement{r00, r10, r20, r30, r40}
-	v.carryPropagate1().carryPropagate2()
+	v.carryPropagate()
 }