--- a/fe_generic.go
+++ b/fe_generic.go
@@ -177,3 +177,18 @@ func feSquareGeneric(v, x *fieldElement) {
 	*v = fieldElement{r00, r10, r20, r30, r40}
 	v.carryPropagate()
 }
+
+// carryPropagate brings the limbs below 52, 51, 51, 51, 51 bits.
+func (v *fieldElement) carryPropagateGeneric() *fieldElement {
+	v.l1 += v.l0 >> 51
+	v.l0 &= maskLow51Bits
+	v.l2 += v.l1 >> 51
+	v.l1 &= maskLow51Bits
+	v.l3 += v.l2 >> 51
+	v.l2 &= maskLow51Bits
+	v.l4 += v.l3 >> 51
+	v.l3 &= maskLow51Bits
+	v.l0 += (v.l4 >> 51) * 19
+	v.l4 &= maskLow51Bits
+	return v
+}