--- a/fe_test.go
+++ b/fe_test.go
@@ -519,3 +519,23 @@ func TestSqrtRatio(t *testing.T) {
 		}
 	}
 }
+
+func TestCarryPropagate(t *testing.T) {
+	asmLikeGeneric := func(a [5]uint64) bool {
+		t1 := &fieldElement{a[0], a[1], a[2], a[3], a[4]}
+		t2 := &fieldElement{a[0], a[1], a[2], a[3], a[4]}
+
+		t1.carryPropagate()
+		t2.carryPropagateGeneric()
+
+		if *t1 != *t2 {
+			t.Logf("got: %#v,\nexpected: %#v", t1, t2)
+		}
+
+		return *t1 == *t2 && isInBounds(t2)
+	}
+
+	if err := quick.Check(asmLikeGeneric, quickCheckConfig1024); err != nil {
+		t.Error(err)
+	}
+}