```
name                    old time/op  new time/op  delta
Add-8                   7.17ns ± 1%  6.43ns ± 1%  -10.29%  (p=0.000 n=10+8)
Mul-8                   26.0ns ± 1%  24.6ns ± 1%   -5.40%  (p=0.000 n=10+10)
Mul32-8                 5.87ns ± 1%  5.87ns ± 1%     ~     (p=0.755 n=10+10)
WideMultCall-8          2.53ns ± 1%  2.54ns ± 0%     ~     (p=0.915 n=10+8)
BasepointMul-8          20.1µs ± 1%  18.7µs ± 1%   -6.94%  (p=0.000 n=9+10)
ScalarMul-8             71.6µs ± 0%  63.9µs ± 1%  -10.82%  (p=0.000 n=8+9)
VartimeDoubleBaseMul-8  68.7µs ± 1%  60.7µs ± 2%  -11.61%  (p=0.000 n=8+9)
MultiscalarMulSize8-8    246µs ± 2%   224µs ± 1%   -8.84%  (p=0.000 n=9+9)
```

@hdevalence @gtank, please check that I'm correct that everything works with 52 bit limbs, especially `reduce`?