@hdevalence 2020-12-11
I haven't looked at the code yet, but I believe that 51-bit limbs can have 3 bits of headroom without problems. There is a detailed derivation and proof that I wrote in the comments of the curve25519-dalek source (in the u64 backend).