@FiloSottile 2020-12-12
Cool, that matches my analysis. I feel like that's enough to merge, but I would still appreciate a quick review of the code!